chore: split utils file
