+++ app/renderer/containers/SessionPage.js
@@ -2,7 +2,7 @@ import {connect} from 'react-redux';
 
 import * as SessionActions from '../actions/Session';
 import Session from '../components/Session/Session';
-import {withTranslation} from '../util';
+import {withTranslation} from '../utils/other';
 
 function mapStateToProps(state) {
   return state.session;
