+++ app/renderer/utils/locator-generation.js
@@ -1,22 +1,8 @@
-import {DOMParser} from '@xmldom/xmldom';
 import _ from 'lodash';
-import {withTranslation as wt} from 'react-i18next';
 import XPath from 'xpath';
 
-import config from '../configs/app.config';
-import {log} from './polyfills';
-
-const VALID_W3C_CAPS = [
-  'platformName',
-  'browserName',
-  'browserVersion',
-  'acceptInsecureCerts',
-  'pageLoadStrategy',
-  'proxy',
-  'setWindowRect',
-  'timeouts',
-  'unhandledPromptBehavior',
-];
+import {log} from '../polyfills';
+import {domParser, findDOMNodeByPath} from './source-parsing';
 
 // Attributes on nodes that are likely to be unique to the node so we should consider first when
 // suggesting xpath locators. These are considered IN ORDER.
@@ -40,8 +26,6 @@ const SIMPLE_STRATEGY_MAPPINGS = [
   ['class', 'class name'],
   ['type', 'class name'],
 ];
-
-export const domParser = new DOMParser();
 
 /**
  * Check whether the provided attribute & value are unique in the source
@@ -120,84 +104,6 @@ export function getSuggestedLocators(selectedElement, sourceXML, isNative) {
   );
   const complexLocators = getComplexSuggestedLocators(selectedElement.path, sourceDoc);
   return _.toPairs({...simpleLocators, ...complexLocators});
-}
-
-/**
- * Get the child nodes of a Node object
- *
- * @param {Node} domNode
- * @returns {Array<Node|null>} list of Nodes
- */
-function childNodesOf(domNode) {
-  if (!domNode || !domNode.hasChildNodes()) {
-    return [];
-  }
-  return _.filter(domNode.childNodes, ['nodeType', domNode.ELEMENT_NODE]);
-}
-
-/**
- * Look up an element in the Document source using the provided path
- *
- * @param {string} path a dot-separated string of indices
- * @param {Document} sourceDoc app source in Document format
- * @returns {Node} element node
- */
-export function findDOMNodeByPath(path, sourceDoc) {
-  let selectedElement = childNodesOf(sourceDoc)[0] || childNodesOf(sourceDoc.documentElement)[0];
-  for (let index of path.split('.')) {
-    selectedElement = childNodesOf(selectedElement)[index];
-  }
-  return selectedElement;
-}
-
-/**
- * Look up an element in the JSON source using the provided path
- *
- * @param {string} path a dot-separated string of indices
- * @param {Object} sourceJSON app source in JSON format
- * @returns {Object} element details in JSON format
- */
-export function findJSONElementByPath(path, sourceJSON) {
-  let selectedElement = sourceJSON;
-  for (let index of path.split('.')) {
-    selectedElement = selectedElement.children[index];
-  }
-  return {...selectedElement};
-}
-
-/**
- * Translates sourceXML to JSON
- *
- * @param {string} sourceXML
- * @returns {Object} source in JSON format
- */
-export function xmlToJSON(sourceXML) {
-  const translateRecursively = (domNode, parentPath = '', index = null) => {
-    const attributes = {};
-    for (let attrIdx = 0; attrIdx < domNode.attributes.length; ++attrIdx) {
-      const attr = domNode.attributes.item(attrIdx);
-      attributes[attr.name] = attr.value;
-    }
-
-    // Dot Separated path of indices
-    const path = _.isNil(index) ? '' : `${!parentPath ? '' : parentPath + '.'}${index}`;
-
-    return {
-      children: childNodesOf(domNode).map((childNode, childIndex) =>
-        translateRecursively(childNode, path, childIndex),
-      ),
-      tagName: domNode.tagName,
-      attributes,
-      path,
-    };
-  };
-  const sourceDoc = domParser.parseFromString(sourceXML);
-  // get the first child element node in the doc. some drivers write their xml differently so we
-  // first try to find an element as a direct descendend of the doc, then look for one in
-  // documentElement
-  const firstChild = childNodesOf(sourceDoc)[0] || childNodesOf(sourceDoc.documentElement)[0];
-
-  return firstChild ? translateRecursively(firstChild) : {};
 }
 
 /**
@@ -519,18 +425,4 @@ export function getOptimalPredicateString(doc, domNode) {
 
     return null;
   }
-}
-
-export function withTranslation(componentCls, ...hocs) {
-  return _.flow(...hocs, wt(config.namespace))(componentCls);
-}
-
-export function addVendorPrefixes(caps) {
-  return caps.map((cap) => {
-    // if we don't have a valid unprefixed cap or a cap with an existing prefix, update it
-    if (!VALID_W3C_CAPS.includes(cap.name) && !_.includes(cap.name, ':')) {
-      cap.name = `appium:${cap.name}`;
-    }
-    return cap;
-  });
 }
